// packages/dler/src/impl/types/mod.ts
export interface AppParams {
	projectName: string;
	cwd: string;
	isDev: boolean;
	memory: ReliverseMemory;
	config: ReliverseConfig;
	mrse: ReliverseConfig[];
	skipPrompts: boolean;
}